+++ example/lib/main.dart
@@ -2,9 +2,73 @@ import 'package:flutter/material.dart';
 import 'package:formx/formx.dart';
 
 void main() {
-  runApp(const MaterialApp(home: Material(child: FormxExample())));
+  runApp(const MaterialApp(home: Material(child: Form2Example())));
 }
 
+class Form2Example extends StatefulWidget {
+  const Form2Example({super.key});
+
+  @override
+  State<Form2Example> createState() => _Form2ExampleState();
+}
+
+class _Form2ExampleState extends State<Form2Example> {
+  var obscure = false;
+
+  void toggleObscure() {
+    setState(() => obscure = !obscure);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Formx(
+      onChanged: print,
+      builder: (context, state, _) {
+        return Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Formx(
+                key: const Key('nested'),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    TextFormField(
+                      key: const Key('name'),
+                    ),
+                    TextFormField(
+                      key: const Key('password'),
+                      obscureText: obscure,
+                      decoration: InputDecoration(
+                        suffixIcon: IconButton(
+                          onPressed: toggleObscure,
+                          icon: Icon(
+                            !obscure ? Icons.remove_red_eye : Icons.security,
+                          ),
+                        ),
+                      ),
+                    ),
+                    TextFormField(
+                      key: const Key('email'),
+                    ),
+                  ],
+                ),
+              ),
+              DropdownButtonFormField(
+                key: const Key('dropdown'),
+                items: const [
+                  DropdownMenuItem(value: 'a', child: Text('Item 1')),
+                  DropdownMenuItem(value: 'b', child: Text('Item 2')),
+                ],
+                onChanged: (_) {},
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
 class User {
   const User({required this.name, required this.email, required this.address});
 
@@ -32,13 +96,6 @@ class FormxExample extends StatelessWidget {
         // Check your console and type, it's alive!
         onChanged: print,
 
-        // Called when all fields are valid and submitted via Enter key or .submit().
-        onSubmitted: (state) {
-          // if (Formx.at(context).validate()) {
-          print('all valid and ready to go: $state');
-          // }
-        },
-
         // Optional initial values for all fields. Tip: Model.toMap() or autofill.
         initialValues: const {
           'name': 'Big',
@@ -49,75 +106,29 @@ class FormxExample extends StatelessWidget {
           },
         },
 
-        // Allows you to modify the field before it's built.
-        fieldModifier: (tag, field) {
-          // `.required()` is an extension. See [TextFormxFieldModifiers]
-          return tag == 'name' ? field.required() : field;
-        },
-
-        // A wrapper that will be applied to all fields. DRY!
-        fieldWrapper: (tag, widget) => Padding(padding: padding, child: widget),
-
-        // [InputDecoration] callback. Tip: scope your form themes!
-        decorationModifier: (tag, decoration) {
-          return decoration?.copyWith(
-            labelText: tag,
-            border: const OutlineInputBorder(),
-            constraints: const BoxConstraints(maxWidth: 300),
-          );
-        },
-
-        // [TextFormField.validator] errorText callback. Tip: nice for i18n key.
-        errorTextModifier: (tag, errorText) {
-          if (tag == 'email') return errorText;
-          return '$errorText.$tag';
-        },
-
         // Your creativity is the limit. Manage your fields however you want.
         // Just make sure there's a [Formx] above them.
-        builder: (context, state) => Scaffold(
+        builder: (context, state, _) => Scaffold(
           body: Center(
             child: Column(
               mainAxisSize: MainAxisSize.min,
               children: [
                 const Text('USER'),
 
-                // Same as TextFormField but copiable and const.
-                const TextFormxField(tag: 'name'),
-
-                /// Stack validators through extensions modifiers.
-                const TextFormxField(tag: 'email')
-                    .validate(
-                        (value) => value.isNotEmpty) // same as .required()
-                    .validate((value) => value.contains('form'))
-                    .validate((value) => value.contains('x'), 'Missing x')
-                    .validate((value) => value.contains('@'), 'Not an email'),
-
-                /// Simple as one line.
-                const TextFormxField(tag: 'password'), // adds suffixIcon
+                // Just add a key to your fields and you're good to go.
+                TextFormField(key: const Key('name')),
+                TextFormField(key: const Key('email')),
+                TextFormField(key: const Key('password')),
 
                 /// Nested fields too!?
                 Formx(
-                  tag: 'address',
-
-                  /// and value modifier? Yes!
-                  valueModifier: (tag, value) {
-                    if (tag == 'number') {
-                      return int.tryParse(value);
-                    }
-                    return value;
-                  },
+                  key: const Key('address'),
                   child: Column(
                     mainAxisSize: MainAxisSize.min,
                     children: [
                       const Text('Address:'),
-
-                      // Use copyWith to create your own extensions!
-                      const TextFormxField(tag: 'street').copyWith(),
-
-                      // Like this cool num validator.
-                      const TextFormxField(tag: 'number')
-                          .validateNum((n) => n < 100),
+                      TextFormField(key: const Key('street')),
+                      TextFormField(key: const Key('number')),
                     ],
                   ),
                 ),
@@ -151,8 +162,8 @@ class FormxExample extends StatelessWidget {
                   // Validate all fields. Just like `Form.validate()`.
                   final isValid = state.validate();
 
-                  // you can also validate specific fields with:
-                  // validate(tag: , tags: , key: , keys: ,);
+                  // you can also validate specific fields:
+                  // validate(['name', 'email']);
 
                   if (isValid) {
                     Map<String, dynamic> form = state.form;
@@ -179,8 +190,8 @@ class ComplexDataStructure extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Formx(
-      initialValues: {
+    return Formx(
+      initialValues: const {
         'user': {
           'name': 'Big',
           'email': 'a@a',
@@ -197,31 +208,31 @@ class ComplexDataStructure extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         children: [
           Formx(
-            tag: 'user',
+            key: const Key('user'),
             child: Column(
               mainAxisSize: MainAxisSize.min,
               children: [
-                TextFormxField(tag: 'name'),
-                TextFormxField(tag: 'email'),
+                TextFormField(key: const Key('name')),
+                TextFormField(key: const Key('email')),
               ],
             ),
           ),
           Formx(
-            tag: 'details',
+            key: const Key('details'),
             child: Column(
               mainAxisSize: MainAxisSize.min,
-              children: <Widget>[
+              children: [
                 Formx(
-                  tag: 'address',
+                  key: const Key('address'),
                   child: Column(
                     mainAxisSize: MainAxisSize.min,
                     children: [
-                      TextFormxField(tag: 'street'),
-                      TextFormxField(tag: 'number'),
+                      TextFormField(key: const Key('street')),
+                      TextFormField(key: const Key('number')),
                     ],
                   ),
                 ),
-                TextFormxField(tag: 'school'),
+                TextFormField(key: const Key('school')),
               ],
             ),
           ),