+++ lib/src/extensions/element_visitors.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 /// A [BuildContext] extension for [Element], [State] and [Widget] visitings.